requirements.txt
snowflake-connector-python==4.1.1
python-dotenv==1.2.1
